<!-- See https://github.com/guardian/recommendations/blob/main/pull-requests.md for recommendations on raising and reviewing pull requests. -->

## What does this change?

Remove join in search query.

Old query:
```
explain analyze SELECT
   fm.id AS i_on_fm,
   fm.external_id AS ei_on_fm,
   fm.ingested_at AS ia_on_fm,
   fm.supplier AS s_on_fm,
   fm.content AS c_on_fm,
   '' AS h_on_fm
FROM fingerpost_wire_entry fm
WHERE NOT EXISTS (
  SELECT 1 FROM fingerpost_wire_entry keywordsExcl
  WHERE fm.id = keywordsExcl.id
    AND (keywordsExcl.content->'keywords') ?| ARRAY['Sports']
)                                             
AND NOT EXISTS (                                                      
  SELECT 1 FROM fingerpost_wire_entry subjectsExcl
  WHERE fm.id = subjectsExcl.id
    AND (subjectsExcl.content->'subjects'->'code') ?| ARRAY[
      'MCC:SPO', 'N2:TEN', 'a1312cat:s', 'iptccat:GXX', 'iptccat:RDR',
      'iptccat:RFC', 'iptccat:RRB', 'iptccat:RRR', 'iptccat:RSF', 'iptccat:RSR',
      'iptccat:SCN', 'iptccat:SCR', 'iptccat:SDA', 'iptccat:SDB', 'iptccat:SDP',
      'iptccat:SDQ', 'iptccat:SDR', 'iptccat:SDS', 'iptccat:SFF', 'iptccat:SJA',
      'iptccat:SJB', 'iptccat:SOD', 'iptccat:SOS', 'iptccat:SRD', 'iptccat:SRN',
      'iptccat:SRR', 'iptccat:SRS', 'iptccat:SRZ', 'iptccat:SSD', 'iptccat:SSF',
      'iptccat:SSO', 'iptccat:SSS', 'iptccat:SST', 'iptccat:STA', 'iptccat:STB',
      'iptccat:STC', 'iptccat:STD', 'iptccat:STR', 'medtop:15000000'
    ]
)
ORDER BY fm.ingested_at DESC
LIMIT 30;
                                QUERY PLAN                                                                                                                                                                                                                                                                                 
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=1.25..429.47 rows=30 width=503) (actual time=0.059..0.851 rows=30 loops=1)
   ->  Nested Loop Anti Join  (cost=1.25..1149330.52 rows=80520 width=503) (actual time=0.058..0.847 rows=30 loops=1)
         ->  Nested Loop Anti Join  (cost=0.83..758059.79 rows=94880 width=471) (actual time=0.046..0.629 rows=34 loops=1)
               ->  Index Scan Backward using ingested_at_index on fingerpost_wire_entry fm  (cost=0.42..246425.52 rows=123967 width=471) (actual time=0.011..0.053 rows=39 loops=1)
               ->  Index Scan using fingerpost_wire_entry_pkey on fingerpost_wire_entry subjectsexcl  (cost=0.42..4.12 rows=1 width=8) (actual time=0.014..0.014 rows=0 loops=39)
                     Index Cond: (id = fm.id)
                     Filter: (((content -> 'subjects'::text) -> 'code'::text) ?| '{MCC:SPO,N2:TEN,a1312cat:s,iptccat:GXX,iptccat:RDR,iptccat:RFC,iptccat:RRB,iptccat:RRR,iptccat:RSF,iptccat:RSR,iptccat:SCN,iptccat:SCR,iptccat:SDA,iptccat:SDB,iptccat:SDP,iptccat:SDQ,iptccat:SDR,iptccat:SDS,iptccat:SFF,iptccat:SJA,iptccat:SJB,iptccat:SOD,iptccat:SOS,iptccat:SRD,iptccat:SRN,iptccat:SRR,iptccat:SRS,iptccat:SRZ,iptccat:SSD,iptccat:SSF,iptccat:SSO,iptccat:SSS,iptccat:SST,iptccat:STA,iptccat:STB,iptccat:STC,iptccat:STD,iptccat:STR,medtop:15000000}'::text[])
                     Rows Removed by Filter: 1
         ->  Index Scan using fingerpost_wire_entry_pkey on fingerpost_wire_entry keywordsexcl  (cost=0.42..4.12 rows=1 width=8) (actual time=0.006..0.006 rows=0 loops=34)
               Index Cond: (id = fm.id)
               Filter: ((content -> 'keywords'::text) ?| '{Sports}'::text[])
               Rows Removed by Filter: 1
 Planning Time: 3.552 ms
 Execution Time: 0.897 ms
(14 rows)
```

New query:
```
explain analyze SELECT
    fm.id AS i_on_fm,
    fm.external_id AS ei_on_fm,
    fm.ingested_at AS ia_on_fm,
    fm.supplier AS s_on_fm,
    fm.content AS c_on_fm,
    '' AS h_on_fm
FROM fingerpost_wire_entry fm
WHERE NOT (       
    (fm.content->'keywords') ?| ARRAY['Sports']
)                              
AND NOT (                                                    
    (fm.content->'subjects'->'code') ?| ARRAY[
      'MCC:SPO', 'N2:TEN', 'a1312cat:s', 'iptccat:GXX', 'iptccat:RDR',
      'iptccat:RFC', 'iptccat:RRB', 'iptccat:RRR', 'iptccat:RSF', 'iptccat:RSR',
      'iptccat:SCN', 'iptccat:SCR', 'iptccat:SDA', 'iptccat:SDB', 'iptccat:SDP',
      'iptccat:SDQ', 'iptccat:SDR', 'iptccat:SDS', 'iptccat:SFF', 'iptccat:SJA',
      'iptccat:SJB', 'iptccat:SOD', 'iptccat:SOS', 'iptccat:SRD', 'iptccat:SRN',
      'iptccat:SRR', 'iptccat:SRS', 'iptccat:SRZ', 'iptccat:SSD', 'iptccat:SSF',
      'iptccat:SSO', 'iptccat:SSS', 'iptccat:SST', 'iptccat:STA', 'iptccat:STB',
      'iptccat:STC', 'iptccat:STD', 'iptccat:STR', 'medtop:15000000'
    ]                                                                           
)                                                                               
ORDER BY fm.ingested_at DESC
LIMIT 30;
                                                              QUERY PLAN                                                                                                                                                                                                                                                                                                               
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.42..87.39 rows=30 width=503) (actual time=0.025..0.697 rows=30 loops=1)
   ->  Index Scan Backward using ingested_at_index on fingerpost_wire_entry fm  (cost=0.42..233422.91 rows=80520 width=503) (actual time=0.024..0.692 rows=30 loops=1)
         Filter: ((NOT ((content -> 'keywords'::text) ?| '{Sports}'::text[])) AND (NOT (((content -> 'subjects'::text) -> 'code'::text) ?| '{MCC:SPO,N2:TEN,a1312cat:s,iptccat:GXX,iptccat:RDR,iptccat:RFC,iptccat:RRB,iptccat:RRR,iptccat:RSF,iptccat:RSR,iptccat:SCN,iptccat:SCR,iptccat:SDA,iptccat:SDB,iptccat:SDP,iptccat:SDQ,iptccat:SDR,iptccat:SDS,iptccat:SFF,iptccat:SJA,iptccat:SJB,iptccat:SOD,iptccat:SOS,iptccat:SRD,iptccat:SRN,iptccat:SRR,iptccat:SRS,iptccat:SRZ,iptccat:SSD,iptccat:SSF,iptccat:SSO,iptccat:SSS,iptccat:SST,iptccat:STA,iptccat:STB,iptccat:STC,iptccat:STD,iptccat:STR,medtop:15000000}'::text[])))
         Rows Removed by Filter: 8
 Planning Time: 1.445 ms
 Execution Time: 0.719 ms
(6 rows)
```

Only apply highlights on final result.

Old query:
```
explain analyze SELECT 
   fm.id as i_on_fm,
   fm.external_id as ei_on_fm,
   fm.ingested_at as ia_on_fm,
   fm.supplier as s_on_fm,
   fm.content as c_on_fm
 , ts_headline('english', fm.content->>'body_text', websearch_to_tsquery('english', 'london')) AS h_on_fm
 FROM fingerpost_wire_entry fm
 WHERE websearch_to_tsquery('english', 'london') @@ fm.combined_textsearch
 ORDER BY fm.ingested_at DESC
 LIMIT 30;
                                                                              QUERY PLAN                                                                               
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.42..1226.80 rows=30 width=503) (actual time=0.705..27.732 rows=30 loops=1)
   ->  Index Scan Backward using ingested_at_index on fingerpost_wire_entry fm  (cost=0.42..233626.28 rows=5715 width=503) (actual time=0.703..27.726 rows=30 loops=1)
         Filter: ('''london'''::tsquery @@ combined_textsearch)
         Rows Removed by Filter: 239
 Planning Time: 0.805 ms
 Execution Time: 27.762 ms
(6 rows)
```

New query:
```
explain analyze WITH search_results AS (
  SELECT 
    fm.id,
    fm.external_id,
    fm.ingested_at,
    fm.supplier,
    fm.content
  FROM fingerpost_wire_entry fm
  WHERE fm.combined_textsearch @@ websearch_to_tsquery('english', 'london')
  ORDER BY fm.ingested_at DESC
  LIMIT 30
)
SELECT 
  id AS i_on_fm,
  external_id AS ei_on_fm,
  ingested_at AS ia_on_fm,
  supplier AS s_on_fm,
  content AS c_on_fm,
  ts_headline('english', content->>'body_text', websearch_to_tsquery('english', 'london')) AS h_on_fm
FROM search_results;
                                                                                 QUERY PLAN                                                                                 
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Subquery Scan on search_results  (cost=0.42..1226.80 rows=30 width=503) (actual time=0.742..27.890 rows=30 loops=1)
   ->  Limit  (cost=0.42..1219.22 rows=30 width=471) (actual time=0.076..1.369 rows=30 loops=1)
         ->  Index Scan Backward using ingested_at_index on fingerpost_wire_entry fm  (cost=0.42..232183.24 rows=5715 width=471) (actual time=0.075..1.364 rows=30 loops=1)
               Filter: (combined_textsearch @@ '''london'''::tsquery)
               Rows Removed by Filter: 243
 Planning Time: 0.194 ms
 Execution Time: 27.919 ms
(7 rows)
```

## How to test

<!-- Provide instructions to help others verify the change. This could take the form of "On PROD, do X and witness Y. On this branch, do X and witness Z. " -->

## How can we measure success?

<!-- Do you expect errors to decrease? Do you expect user journeys to be simplified? What can be used to prove this? A filtered view of logs or analytics, etc? -->

## Have we considered potential risks?

<!-- What are the potential risks and how can they be mitigated? Does an error require an alarm? Should user help, infosec, or legal be informed of this change? Is private information guarded? Do we need to add anything in the backlog? -->

## Images

<!-- Usually only applicable to UI changes, what did it look like before and what will it look like after? -->

## Accessibility

<!-- Usually only applicable to UI changes, check the boxes if you are satisfied that your changes pass these tests -->

-   [ ] [Tested with screen reader](https://github.com/guardian/accessibility/blob/main/people-and-technology/03-visual.md#screen-reader)
-   [ ] [Navigable with keyboard](https://github.com/guardian/accessibility/blob/main/people-and-technology/02-physical.md#keyboard)
-   [ ] [Colour contrast passed](https://github.com/guardian/accessibility/blob/main/people-and-technology/03-visual.md#contrast)
-   [ ] [The change doesn't use only colour to convey meaning](https://github.com/guardian/accessibility/blob/main/people-and-technology/03-visual.md#use-of-colour)