@sb-dev 2025-02-10
Closing the PR as performance didn't improve as expected.